Separate path/file operations into separate classes and change to be real objects instead of static methods
