+++ src/main/java/httpserver/file/Html.java
@@ -1,13 +1,11 @@
-package httpserver.fileutils;
+package httpserver.file;
 
 import java.nio.file.Path;
-
-import static httpserver.fileutils.Files.isDir;
 
 public class Html {
     public static String hrefString(Path root, Path fullFilePath) {
         Path diff = root.relativize(fullFilePath);
-        if (isDir(fullFilePath)) {
+        if (new PathExaminer().isDir(fullFilePath)) {
             return diff.toString() + "/";
         }
         return diff.toString();
