+++ src/test/java/httpserver/responder/GetResponderTest.java
@@ -6,11 +6,10 @@ import org.junit.Test;
 
 import java.io.IOException;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.HashMap;
 
-import static httpserver.fileutils.FileHelpers.tempDir;
-import static httpserver.fileutils.FileHelpers.tempFileOptions;
+import static httpserver.file.FileHelpers.tempDir;
+import static httpserver.file.FileHelpers.tempFileOptions;
 import static java.nio.file.Files.write;
 import static junit.framework.TestCase.*;
 import static org.junit.Assert.assertEquals;
